Restructure InteractionLayer to a separate component
