+++ src/layers/interaction/index.ts
@@ -1,5 +1,2 @@
-export * from './InteractionHandling'
 export * from './model'
-export * from './MouseCursor'
-export * from './CursorLabel'
-export * from './MouseAwareSvg'
+export * from './InteractionLayer'
