+++ src/layers/interaction/trimmer/index.ts
